fix:template

+++ AT_SIMULATION/src/schema/visio.py
@@ -69,11 +69,25 @@ class Node(Base):
 
 
 def delete_associated_node(mapper, connection: Connection, target):
-    connection.execute(
-        Node.__table__.delete()
+    node_ids = connection.execute(
+        Node.__table__.select()
+        .with_only_columns(Node.id)  
         .where(Node.object_table == target.__tablename__)
         .where(Node.object_id == target.id)
-    )
+    ).fetchall()
+
+    node_ids = [row[0] for row in node_ids]
+
+    if node_ids:
+        connection.execute(
+            Edge.__table__.delete()
+            .where(Edge.from_node.in_(node_ids) | Edge.to_node.in_(node_ids))
+        )
+
+        connection.execute(
+            Node.__table__.delete()
+            .where(Node.id.in_(node_ids))
+        )
 
 
 event.listen(Resource, "after_delete", delete_associated_node)
